# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Quiz Application API for recruitment
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quizzes:
    get:
      operationId: listQuizzes
      tags: [quizzes]
      summary: List all quizzes
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
    post:
      operationId: createQuiz
      tags: [quizzes]
      summary: Create a new quiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{id}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      summary: Get quiz by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
        "404":
          description: Not found
    patch:
      operationId: updateQuiz
      tags: [quizzes]
      summary: Update quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
    delete:
      operationId: deleteQuiz
      tags: [quizzes]
      summary: Delete quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /quizzes/{quizId}/questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List questions for a quiz
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create a question
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /quizzes/{quizId}/questions/{id}:
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a question
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Quiz:
      type: object
      required: [id, title, description, category, durationMinutes, passingScore, status, questionCount, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        durationMinutes:
          type: integer
        passingScore:
          type: integer
        status:
          type: string
          enum: [draft, active, archived]
        questionCount:
          type: integer
        createdAt:
          type: string
    QuizInput:
      type: object
      required: [title, description, category, durationMinutes, passingScore]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        durationMinutes:
          type: integer
        passingScore:
          type: integer
        status:
          type: string
          enum: [draft, active, archived]
    QuizUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        durationMinutes:
          type: integer
        passingScore:
          type: integer
        status:
          type: string
          enum: [draft, active, archived]
    Question:
      type: object
      required: [id, quizId, text, type, points, orderIndex, options, correctAnswer]
      properties:
        id:
          type: integer
        quizId:
          type: integer
        text:
          type: string
        type:
          type: string
          enum: [multiple_choice, true_false, short_answer]
        points:
          type: integer
        orderIndex:
          type: integer
        options:
          type: ["array", "null"]
          items:
            type: string
        correctAnswer:
          type: string
    QuestionInput:
      type: object
      required: [text, type, points, correctAnswer]
      properties:
        text:
          type: string
        type:
          type: string
          enum: [multiple_choice, true_false, short_answer]
        points:
          type: integer
        orderIndex:
          type: integer
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
    QuestionUpdate:
      type: object
      properties:
        text:
          type: string
        type:
          type: string
          enum: [multiple_choice, true_false, short_answer]
        points:
          type: integer
        orderIndex:
          type: integer
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
...