# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Quiz Application API for recruitment
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionType } from './questionType';

export interface Question {
  id: number;
  quizId: number;
  text: string;
  type: QuestionType;
  points: number;
  orderIndex: number;
  /** @nullable */
  options: string[] | null;
  correctAnswer: string;
}