# lib/api-zod/src/generated/types/questionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Quiz Application API for recruitment
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionInputType } from './questionInputType';

export interface QuestionInput {
  text: string;
  type: QuestionInputType;
  points: number;
  orderIndex?: number;
  options?: string[];
  correctAnswer: string;
}