# lib/api-zod/src/generated/types/questionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Quiz Application API for recruitment
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionUpdateType } from './questionUpdateType';

export interface QuestionUpdate {
  text?: string;
  type?: QuestionUpdateType;
  points?: number;
  orderIndex?: number;
  options?: string[];
  correctAnswer?: string;
}