# lib/api-zod/src/generated/types/quizInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Quiz Application API for recruitment
 * OpenAPI spec version: 0.1.0
 */
import type { QuizInputStatus } from './quizInputStatus';

export interface QuizInput {
  title: string;
  description: string;
  category: string;
  durationMinutes: number;
  passingScore: number;
  status?: QuizInputStatus;
}