# lib/api-zod/src/generated/types/quizUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Quiz Application API for recruitment
 * OpenAPI spec version: 0.1.0
 */
import type { QuizUpdateStatus } from './quizUpdateStatus';

export interface QuizUpdate {
  title?: string;
  description?: string;
  category?: string;
  durationMinutes?: number;
  passingScore?: number;
  status?: QuizUpdateStatus;
}